+++ src/backend/langboard/core/bootstrap/uvicorn/supervisors/multiprocess.py
@@ -42,6 +42,7 @@ from socket import socket
 from typing import Any, Callable
 import click
 from uvicorn.config import Config
+from .....Constants import ENVIRONMENT
 from ....broadcast import DispatcherQueue
 from .._subprocess import get_subprocess, run_broker
 
@@ -159,18 +160,24 @@ class Multiprocess:
             process = Process(self.config, self.target, self.sockets)
             process.start()
             self.processes.append(process)
-        self.broker_process = run_broker(is_restarting=False)
+
+        if ENVIRONMENT == "local":
+            self.broker_process = run_broker(is_restarting=False)
 
     def terminate_all(self) -> None:
         for process in self.processes:
             process.terminate()
-        self.broker_process.terminate()
-        self.broker_process.kill()
+
+        if ENVIRONMENT == "local":
+            self.broker_process.terminate()
+            self.broker_process.kill()
 
     def join_all(self) -> None:
         for process in self.processes:
             process.join()
-        self.broker_process.join()
+
+        if ENVIRONMENT == "local":
+            self.broker_process.join()
 
     def restart_all(self) -> None:
         for idx, process in enumerate(self.processes):
@@ -179,10 +186,12 @@ class Multiprocess:
             new_process = Process(self.config, self.target, self.sockets)
             new_process.start()
             self.processes[idx] = new_process
-        self.broker_process.terminate()
-        self.broker_process.kill()
-        self.broker_process.join()
-        self.broker_process = run_broker(is_restarting=True)
+
+        if ENVIRONMENT == "local":
+            self.broker_process.terminate()
+            self.broker_process.kill()
+            self.broker_process.join()
+            self.broker_process = run_broker(is_restarting=True)
 
     def run(self) -> None:
         message = f"Started parent process [{os.getpid()}]"
@@ -221,11 +230,12 @@ class Multiprocess:
             process.start()
             self.processes[idx] = process
 
-        if not self.broker_process.is_alive():
-            self.broker_process.terminate()
-            self.broker_process.kill()
-            self.broker_process.join()
-            self.broker_process = run_broker(is_restarting=True)
+        if ENVIRONMENT == "local":
+            if not self.broker_process.is_alive():
+                self.broker_process.terminate()
+                self.broker_process.kill()
+                self.broker_process.join()
+                self.broker_process = run_broker(is_restarting=True)
 
     def handle_signals(self) -> None:
         for sig in tuple(self.signal_queue):
